refactor: code-quality cleanup across backend and frontend
Surgical cleanup driven by an 8-agent audit; every change re-verified and
gated on `pytest -m "not integration"` (1424 pass), ruff, and the Vite build.

Dead code removed (verified zero callers):
- SimulationIPCServer, GraphBuilderService.build_graph_async/_build_graph_worker,
  dkg/waybackclaw health_check + probe constants, GraphMemoryUpdater.get_all_stats
- frontend: 6 unused axios wrappers, legacy toggleLocale, orphaned entity-card CSS

DRY / structure:
- new leaf simulation_run_state.py (RunnerStatus/AgentAction/RoundSummary/
  SimulationRunState) drops 5 TYPE_CHECKING notify/runner import shims (no cycle)
- new leaf _notify_base.py: shared per-channel Dedup + parameterized post_json
- _fan_out_notifications collapses the 3x notify dispatch in simulation_runner
- bucket_snapshots centralizes the belief stance-bucketing inlined in 5 services
  (behavior-identical; validated by the stat-builder unit tests)

Error handling (task: no error hiding) — surface internal bugs, keep real
external defense:
- narrowed dead `except (X, Exception)` tuples and redundant `except Exception:
  pass` (simulation.py, observability.py, graph_tools.py)
- removed two `pdb.set_trace()` debugger breakpoints in recsys.py

Types: provably-wrong `Dict[str, str]` annotations -> TypedDicts (ResolvedPaths,
RunInstructions, SavedFileInfo); `Optional[callable]` -> `Optional[Callable[...]]`.

Net ~680 fewer lines of logic. No behavior change intended; two pre-existing
bugs (Project.chunk_size 3-way default mismatch, repro_export director-events
filename) were left for a maintainer decision.
