+++ backend/app/services/__init__.py
@@ -16,12 +16,12 @@ from .simulation_config_generator import (
     EventConfig,
     PlatformConfig
 )
-from .simulation_runner import (
-    SimulationRunner,
+from .simulation_runner import SimulationRunner
+from .simulation_run_state import (
     SimulationRunState,
     RunnerStatus,
     AgentAction,
-    RoundSummary
+    RoundSummary,
 )
 from .graph_memory_updater import (
     GraphMemoryUpdater,
@@ -30,7 +30,6 @@ from .graph_memory_updater import (
 )
 from .simulation_ipc import (
     SimulationIPCClient,
-    SimulationIPCServer,
     IPCCommand,
     IPCResponse,
     CommandType,
@@ -64,7 +63,6 @@ __all__ = [
     'GraphMemoryManager',
     'AgentActivity',
     'SimulationIPCClient',
-    'SimulationIPCServer',
     'IPCCommand',
     'IPCResponse',
     'CommandType',
