+++ backend/app/services/email_notify.py
@@ -70,13 +70,11 @@ import threading
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 from email.utils import formataddr
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 from ..utils.logger import get_logger
-
-if TYPE_CHECKING:
-    from .simulation_runner import SimulationRunState
-
+from .simulation_run_state import SimulationRunState
+from ._notify_base import Dedup
 
 logger = get_logger("miroshark.email_notify")
 
@@ -122,26 +120,12 @@ COLOR_FAILED = "#f59e0b"
 
 
 # Per-process dedup — same shape as the Discord / Slack notifiers.
-_FIRED: set[Tuple[str, str]] = set()
-_FIRED_LOCK = threading.Lock()
-_FIRED_MAX = 4096
-
-
-def _mark_fired(sim_id: str, status: str) -> bool:
-    key = (sim_id, status)
-    with _FIRED_LOCK:
-        if key in _FIRED:
-            return False
-        if len(_FIRED) >= _FIRED_MAX:
-            _FIRED.pop()
-        _FIRED.add(key)
-        return True
+_dedup = Dedup()
 
 
 def reset_dedup_for_tests() -> None:
     """Clear the in-process dedup set. Test-only convenience."""
-    with _FIRED_LOCK:
-        _FIRED.clear()
+    _dedup.reset()
 
 
 # ── env-var resolution ────────────────────────────────────────────────
@@ -693,7 +677,7 @@ def notify_if_configured(
     if not to_addrs:
         return
 
-    if not _mark_fired(simulation_id, status):
+    if not _dedup.mark(simulation_id, status):
         return
 
     from . import webhook_service
