+++ backend/app/services/slack_notify.py
@@ -48,17 +48,13 @@ Message shape (Block Kit ``blocks``)::
 
 from __future__ import annotations
 
-import json
 import os
 import threading
-import urllib.error
-import urllib.request
-from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple
+from typing import Any, Dict, Optional, Tuple
 
 from ..utils.logger import get_logger
-
-if TYPE_CHECKING:
-    from .simulation_runner import SimulationRunState
+from .simulation_run_state import SimulationRunState
+from ._notify_base import Dedup, post_json
 
 logger = get_logger('miroshark.slack_notify')
 
@@ -78,26 +74,12 @@ BAR_EMPTY = "░"
 BAR_WIDTH = 10
 
 
-_FIRED: set[Tuple[str, str]] = set()
-_FIRED_LOCK = threading.Lock()
-_FIRED_MAX = 4096
-
-
-def _mark_fired(sim_id: str, status: str) -> bool:
-    key = (sim_id, status)
-    with _FIRED_LOCK:
-        if key in _FIRED:
-            return False
-        if len(_FIRED) >= _FIRED_MAX:
-            _FIRED.pop()
-        _FIRED.add(key)
-        return True
+_dedup = Dedup()
 
 
 def reset_dedup_for_tests() -> None:
     """Clear the in-process dedup set. Test-only convenience."""
-    with _FIRED_LOCK:
-        _FIRED.clear()
+    _dedup.reset()
 
 
 def _resolve_webhook_url() -> str:
@@ -289,29 +271,7 @@ def build_slack_message(payload: Dict[str, Any]) -> Dict[str, Any]:
 
 def _post_json(url: str, body: Dict[str, Any], timeout: float) -> Tuple[bool, str]:
     """Issue the POST. Returns ``(ok, message)`` — never raises."""
-    try:
-        encoded = json.dumps(body).encode("utf-8")
-    except Exception as exc:
-        return False, f"Could not serialize Slack payload: {exc}"
-
-    headers = {
-        "Content-Type": "application/json; charset=utf-8",
-        "User-Agent": SLACK_USER_AGENT,
-    }
-    req = urllib.request.Request(url, data=encoded, method="POST", headers=headers)
-    try:
-        with urllib.request.urlopen(req, timeout=timeout) as resp:
-            code = resp.getcode()
-            if 200 <= code < 300:
-                return True, f"HTTP {code}"
-            return False, f"HTTP {code}"
-    except urllib.error.HTTPError as exc:
-        return False, f"HTTP {exc.code}"
-    except urllib.error.URLError as exc:
-        reason = getattr(exc, "reason", exc)
-        return False, f"URL error: {reason}"
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
+    return post_json(url, body, timeout, user_agent=SLACK_USER_AGENT, label="Slack")
 
 
 def send_slack_payload(url: str, message: Dict[str, Any]) -> Tuple[bool, str]:
@@ -361,7 +321,7 @@ def notify_if_configured(
     if not url:
         return
 
-    if not _mark_fired(simulation_id, status):
+    if not _dedup.mark(simulation_id, status):
         return
 
     from . import webhook_service
